+++ packages/localization/package.json
@@ -27,9 +27,9 @@
     "prepublishOnly": "npm run clean && npm run build"
   },
   "devDependencies": {
-    "@openui5/sap.ui.core": "1.101.0",
+    "@openui5/sap.ui.core": "1.109.0",
     "@ui5/webcomponents-tools": "1.7.4",
-    "chromedriver": "104.0.0",
+    "chromedriver": "107.0.3",
     "mkdirp": "^1.0.4",
     "resolve": "^1.20.0"
   },
