+++ packages/theming/package.json
@@ -33,7 +33,7 @@
   },
   "devDependencies": {
     "@ui5/webcomponents-tools": "1.7.4",
-    "chromedriver": "104.0.0",
+    "chromedriver": "107.0.3",
     "cssnano": "^4.1.11",
     "globby": "^13.1.1",
     "json-beautify": "^1.1.1",
